silx.gui.data.NXdataWidgets: Move ArrayCurvePlot in its own file
